pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autoqa"
version = "0.1.0"
description = "Automated web testing system using BrowserUse"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "langchain-openai>=0.0.5",
    "langchain-google-genai>=0.0.3",
    "browser-use>=0.1.41",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.setuptools]
packages = ["autoqa"]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88
